restructure repository to move documentation into component folders
