+++ api/index.js
@@ -1,0 +1,5 @@
+export {
+	registerBlockExtension,
+	// continue to export misspelled version of api for backwards compatibility
+	registerBlockExtension as registerBlockExtention,
+} from './registerBlockExtension';
