+++ components/clipboard-button/index.js
