+++ components/content-picker/PickedItem.js
