+++ components/content-picker/index.js
