+++ components/icon-picker/icon-picker-toolbar-button.js
