+++ components/icon-picker/icon-picker.js
