+++ components/icon-picker/icon.js
