+++ components/icon-picker/index.js
