+++ components/icon-picker/inline-icon-picker.js
