+++ components/index.js
@@ -1,0 +1,9 @@
+export { IsAdmin } from './is-admin';
+export { Optional } from './optional';
+export { InnerBlockSlider } from './inner-block-slider';
+export { IconPicker, Icon, IconPickerToolbarButton, InlineIconPicker } from './icon-picker';
+export { CustomBlockAppender } from './custom-block-appender';
+export { ContentSearch } from './content-search';
+export { ContentPicker } from './content-picker';
+export { ColorSetting } from './color-settings';
+export { ClipboardButton } from './clipboard-button';
