+++ components/inner-block-slider/icons.js
