+++ components/inner-block-slider/index.js
