+++ components/is-admin/index.js
