+++ icons/index.js
@@ -1,0 +1,2 @@
+export { registerIcons } from './api/register';
+export { store as iconStore } from './store';
