+++ index.js
@@ -1,24 +1,4 @@
-export { ContentPicker } from './components/ContentPicker';
-export { ContentSearch } from './components/ContentSearch';
-export { InnerBlockSlider } from './components/InnerBlockSlider';
-export { ClipboardButton } from './components/ClipboardButton';
-export { IsAdmin } from './components/is-admin';
-export { useHasSelectedInnerBlock } from './hooks/use-has-selected-inner-block';
-export { useRequestData } from './hooks/use-request-data';
-export { default as CustomBlockAppender } from './components/CustomBlockAppender';
-export {
-	registerBlockExtension,
-	// continue to export misspelled version of api for backwards compatibility
-	registerBlockExtension as registerBlockExtention,
-} from './api/registerBlockExtension';
-export { useIcons, useIcon } from './hooks/use-icons';
-export { useFilteredList } from './hooks/use-filtered-list';
-export {
-	IconPicker,
-	Icon,
-	IconPickerToolbarButton,
-	InlineIconPicker,
-} from './components/IconPicker';
-export { registerIcons } from './icons/api/register';
-export { store as iconStore } from './icons/store';
-export { ColorSetting } from './components/ColorSetting';
+export * from './api';
+export * from './icons';
+export * from './hooks';
+export * from './components';
